Add files via upload
This design fits RP2040 Breakout Boards, including the advanced breakout board and basic breakout board. The boards have the same footprint as many Brook products, meaning that this design will also fit several commercially available PCBs for arcade controllers.

The standoff posts were designed to accommodate M3x12 screws or shorter, and affixed to the inside of an enclosure with double-sided tape or hot glue. 

If an M3x16 or longer screw is used, it will protrude through the mounting posts and could be affixed to an enclosure that way if desired.